+++ src/components/FormList/index.ts
@@ -1,0 +1,1 @@
+export {default as FormList} from './FormList';
