+++ src/components/index.ts
@@ -3,6 +3,7 @@ export {ApiInfo, ApisList} from './Apis';
 export {Dashboard} from './Dashboard';
 export {EnvoyFleetInfoModal} from './EnvoyFleetInfoModal';
 export {EnvoyFleetInfo, EnvoyFleetsList} from './EnvoyFleets';
+export {FormList} from './FormList';
 export {InfoPaneCRD} from './InfoPaneCRD';
 export {InfoTabs} from './InfoTabs';
 export {KuskExtensions} from './KuskExtensions';
